+++ osu.Game.Tournament/Screens/Ladder/Components/DrawableTournamentRound.cs
@@ -47,7 +47,7 @@ namespace osu.Game.Tournament.Screens.Ladder.Components
             };
 
             name = round.Name.GetBoundCopy();
-            name.BindValueChanged(n => textName.Text = ((losers ? "Losers " : "") + round.Name).ToUpper(), true);
+            name.BindValueChanged(n => textName.Text = ((losers ? "输者晋级赛 " : "") + round.Name).ToUpper(), true);
 
             description = round.Description.GetBoundCopy();
             description.BindValueChanged(n => textDescription.Text = round.Description.Value?.ToUpper(), true);